requirements.txt
langgraph==0.3.31
langchain-core==0.1.54
openai==1.0.0
python-dotenv==1.0.0 